+++ src/core/transaction.ts
@@ -1,13 +1,14 @@
-import { BoxDBError } from './errors';
-
 import {
-  IDBData,
-  BoxSchema,
-  TransactionTask,
-  CursorTransactionTask,
+  Schema,
+  Data,
+  Task,
+  Transaction,
+  TransactionResult,
   TransactionType,
-  TransactionMode,
+  IDBData,
 } from '../types';
+import { BoxDBError } from './errors';
+import { OptionalData } from '../types/data';
 
 const READONLY_TYPES = [
   TransactionType.GET,
@@ -23,7 +24,51 @@ const HAS_VALUE_TYPES = [
   TransactionType.ADD,
 ];
 
-export default class BoxTransaction {
+enum BoxTransactionMode {
+  READ = 'readonly',
+  WRITE = 'readwrite',
+}
+
+export enum CursorDirection {
+  ASC = 'next',
+  ASC_UNIQUE = 'nextunique',
+  DESC = 'prev',
+  DESC_UNIQUE = 'prevunique',
+}
+
+export type BoxRange<S extends Schema> = {
+  index?: Extract<keyof S, string>;
+  value: IDBKeyRange;
+};
+
+export type FilterFunction<S extends Schema> = (value: Data<S>) => boolean;
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type IDBArgument = [any, any?];
+
+export type TaskParameter<S extends Schema> = {
+  args?: IDBArgument;
+  direction?: CursorDirection | null;
+  filter?: FilterFunction<S>[];
+  range?: BoxRange<S> | null;
+  limit?: number;
+  updateValue?: OptionalData<S>;
+};
+
+export interface BoxTask extends Task<IDBArgument> {
+  name: string;
+}
+
+export interface BoxCursorTask<S extends Schema> extends BoxTask {
+  direction?: CursorDirection | null;
+  filter?: FilterFunction<S>[];
+  range?: BoxRange<S> | null;
+  target?: IDBKeyPath;
+  limit?: number;
+  updateValue?: Partial<Data<S>>;
+}
+
+export default class BoxTransaction implements Transaction<BoxTask> {
   private idb: {
     value: IDBDatabase | null;
   } = { value: null };
@@ -49,9 +94,7 @@ export default class BoxTransaction {
    *
    * @param tasks Transaction tasks
    */
-  run<S extends BoxSchema, T extends TransactionTask>(
-    ...tasks: T[]
-  ): Promise<void | IDBData | IDBData[]> {
+  run<S extends Schema>(...tasks: BoxTask[]): Promise<TransactionResult<S>> {
     const db = this.idb.value;
     if (db === null) {
       throw new BoxDBError('Database not ready');
@@ -72,8 +115,8 @@ export default class BoxTransaction {
 
     // Select transaction mode
     const mode = tasks.some((task) => !READONLY_TYPES.includes(task.type))
-      ? TransactionMode.WRITE
-      : TransactionMode.READ;
+      ? BoxTransactionMode.WRITE
+      : BoxTransactionMode.READ;
 
     return new Promise((resolve, reject) => {
       // Open new transaction
@@ -96,7 +139,7 @@ export default class BoxTransaction {
             action === TransactionType.$DELETE
           ) {
             // using cursor
-            this.cursor<S>(objectStore, task as CursorTransactionTask<S>).then(
+            this.cursor<S>(objectStore, task as BoxCursorTask<S>).then(
               (records) => (res = records),
             );
           } else {
@@ -127,9 +170,9 @@ export default class BoxTransaction {
    * @param objectStore Object store object
    * @param task Current task
    */
-  private cursor<S extends BoxSchema>(
+  private cursor<S extends Schema>(
     objectStore: IDBObjectStore,
-    task: CursorTransactionTask<S>,
+    task: BoxCursorTask<S>,
   ): Promise<void | IDBData | IDBData[]> {
     const filter = task.filter;
     const range = task.range;