+++ src/types/schema.ts
@@ -1,0 +1,16 @@
+import { DataType } from './data';
+
+export type ConfiguredType = {
+  type: DataType;
+  key?: boolean;
+  index?: boolean;
+  unique?: boolean;
+};
+
+export type ConfiguredSchema = {
+  [field: string]: ConfiguredType;
+};
+
+export type Schema = {
+  readonly [field: string]: ConfiguredType | DataType;
+};